requirements.txt
streamlit~=1.52.1
pydub~=0.25.1
edge-tts~=7.2.7
pymupdf~=1.26.7
EbookLib>=0.18
beautifulsoup4>=4.12.0
python-docx>=1.1.0
openai>=1.10.0
python-dotenv>=1.0.0
tenacity>=8.2.0
